feat : Make project modal components
