+++ wise-office-client/src/components/modal/index.ts
@@ -2,6 +2,6 @@ export { default as ConfirmModal } from "./ConfirmModal";
 
 export { default as LogWriteModal } from "./LogWriteModal";
 
-export { default as ProjectModal } from "./ProjectModal";
+export { default as ProjectModal } from "./ProjectModal/ProjectModal";
 
 export { default as DeleteModal } from "./DeleteModal";
